// types/mentor.ts
export type Mentor = {
  id: number;
  name: string;
  title: string;         // fun title like "Scholarship Ninja"
  designation: string;   // current job or status
  motto: string;         // small motivational quote
  image: string;         
  graduateFrom: string;  // where they studied
  achievements: string[]; // list of achievements
};